backend/requirements.txt
# backend/requirements.txt
Flask
psycopg2-binary

